refactor: unify Social type at types/Social.ts, drop social_id + Supabase import
